src/python_backend/requirements.txt
fastapi
uvicorn[standard]
websockets
opencv-python
numpy
aiohttp
pillow
python-dotenv
openai
requests
stripe
supabase
